Clean up semantics, allow setting cafe explictly, or via lb api
