+++ DEVELOPMENT.md
@@ -48,7 +48,7 @@ You should store a file called, `.env` in the root of your Textile Mobile folder
 
 ```
 RN_TEXTILE_CAFE_OVERRIDE="{optional-cafe-url}"
-RN_TEXTILE_CAFE_GATEWAY_URL="http://{gateway-address}"
+RN_TEXTILE_GATEWAY_URL="http://{gateway-address}"
 RN_TEXTILE_CAFE_API_PIN_PATH="/cafe/v0/pin"
 RN_TEXTILE_CAMERA_ROLL_THREAD_KEY="photos-camera_roll"
 RN_TEMPORARY_REFERRAL="SOMETHING"
